Rename configuration class for that creates `TextEncryptorLocator` and make the autowired encryptor non mandatory
