+++ spring-cloud-config-server/src/test/java/org/springframework/cloud/config/server/encryption/SymmetricEncryptionIntegrationTests.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.cloud.config.server.ConfigServerApplication;
+import org.springframework.cloud.config.server.config.SymmetricKeyEncryptorLocatorConfiguration;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -16,7 +17,7 @@ import static org.junit.Assert.assertEquals;
 public class SymmetricEncryptionIntegrationTests {
 
 	@RunWith(SpringRunner.class)
-	@SpringBootTest(classes = {ConfigServerApplication.class,SymmetricKeyEncryptor.class},
+	@SpringBootTest(classes = {ConfigServerApplication.class, SymmetricKeyEncryptorLocatorConfiguration.class},
 			webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 	@ActiveProfiles({"test", "native"})
 	public static class SpringAppJsonConfigSymmetricEncryptionIntegrationTests {
@@ -37,7 +38,7 @@ public class SymmetricEncryptionIntegrationTests {
 	}
 
 	@RunWith(SpringRunner.class)
-	@SpringBootTest(classes = {ConfigServerApplication.class,SymmetricKeyEncryptor.class},
+	@SpringBootTest(classes = {ConfigServerApplication.class, SymmetricKeyEncryptorLocatorConfiguration.class},
 		properties = "spring.cloud.bootstrap.name:symmetric-key-bootstrap",
 			webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 	@ActiveProfiles({"test", "native"})
@@ -53,20 +54,4 @@ public class SymmetricEncryptionIntegrationTests {
 		}
 	}
 
-	@RunWith(SpringRunner.class)
-	@SpringBootTest(classes = {ConfigServerApplication.class,SymmetricKeyEncryptor.class},
-		properties = "spring.config.name:symmetric-key-bootstrap",
-			webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-	@ActiveProfiles({"test", "native"})
-	public static class ApplicationConfigSymmetricEncryptionIntegrationTests {
-
-		@Autowired
-		private TestRestTemplate testRestTemplate;
-
-		@Test
-		public void symmetricEncryptionCannotBeConfiguredInApplicationContext() throws Exception {
-			ResponseEntity<String> entity = testRestTemplate.getForEntity("/encrypt/status", String.class);
-			assertEquals(entity.getStatusCode().value(), 404);
-		}
-	}
 }
